+++ src/main/java/me/banbeucmas/oregen3/utils/hooks/uSkyBlockHook.java
@@ -15,7 +15,7 @@ public class uSkyBlockHook implements SkyblockHook {
     }
 
     @Override
-    public long getIslandLevel(final UUID uuid, final Location loc) {
+    public double getIslandLevel(final UUID uuid, final Location loc) {
         return (long) usb.getIslandLevel(Bukkit.getPlayer(uuid));
     }
